+++ core/data/src/main/java/com/peekr/core/data/file/network/FileNetworkDataSource.kt
@@ -1,7 +1,7 @@
 package com.peekr.core.data.file.network
 
 import com.peekr.core.common.logger.AppLogger
-import com.peekr.core.data.file.response.PresignedUrlResponse
+import com.peekr.core.data.file.network.response.PresignedUrlResponse
 import com.peekr.core.data.network.DefaultOkHttpClient
 import com.peekr.core.data.network.util.NetworkErrorType
 import com.peekr.core.data.network.util.NetworkResult
