+++ core/data/src/main/java/com/peekr/core/data/keyword/network/KeywordNetworkDataSource.kt
@@ -1,7 +1,7 @@
 package com.peekr.core.data.keyword.network
 
-import com.peekr.core.data.keyword.model.request.CreateKeywordRequest
-import com.peekr.core.data.keyword.model.response.KeywordResponse
+import com.peekr.core.data.keyword.network.request.CreateKeywordRequest
+import com.peekr.core.data.keyword.network.response.KeywordResponse
 import com.peekr.core.data.network.util.NetworkResult
 import com.peekr.core.data.network.util.networkCall
 import com.peekr.core.domain.model.KeywordId
