+++ core/data/src/main/java/com/peekr/core/data/keyword/repository/KeywordRepositoryImpl.kt
@@ -1,9 +1,9 @@
 package com.peekr.core.data.keyword.repository
 
 import com.peekr.core.common.IO
-import com.peekr.core.data.keyword.model.request.CreateKeywordRequest
-import com.peekr.core.data.keyword.model.response.toDomainModel
 import com.peekr.core.data.keyword.network.KeywordDataSource
+import com.peekr.core.data.keyword.network.request.CreateKeywordRequest
+import com.peekr.core.data.keyword.network.response.toDomainModel
 import com.peekr.core.data.network.util.NetworkResult
 import com.peekr.core.data.network.util.toErrorType
 import com.peekr.core.domain.coroutine.safeResultFlow
