+++ core/data/src/main/java/com/peekr/core/data/userKeyword/network/UserKeywordDataSource.kt
@@ -1,10 +1,10 @@
 package com.peekr.core.data.userKeyword.network
 
 import com.peekr.core.data.network.util.NetworkResult
-import com.peekr.core.data.userKeyword.model.request.CreateUserKeywordRequest
-import com.peekr.core.data.userKeyword.model.request.PatchUserKeywordRequest
-import com.peekr.core.data.userKeyword.model.response.UserKeywordResponse
-import com.peekr.core.data.userKeyword.model.response.UserKeywordsResponse
+import com.peekr.core.data.userKeyword.network.request.CreateUserKeywordRequest
+import com.peekr.core.data.userKeyword.network.request.PatchUserKeywordRequest
+import com.peekr.core.data.userKeyword.network.response.UserKeywordResponse
+import com.peekr.core.data.userKeyword.network.response.UserKeywordsResponse
 import com.peekr.core.domain.model.UserId
 import com.peekr.core.domain.model.UserKeywordId
