+++ core/data/src/main/java/com/peekr/core/data/userKeyword/network/UserKeywordNetworkDataSource.kt
@@ -3,10 +3,10 @@ package com.peekr.core.data.userKeyword.network
 import com.peekr.core.data.network.util.NetworkResult
 import com.peekr.core.data.network.util.networkCall
 import com.peekr.core.data.network.util.networkCallWithoutResponse
-import com.peekr.core.data.userKeyword.model.request.CreateUserKeywordRequest
-import com.peekr.core.data.userKeyword.model.request.PatchUserKeywordRequest
-import com.peekr.core.data.userKeyword.model.response.UserKeywordResponse
-import com.peekr.core.data.userKeyword.model.response.UserKeywordsResponse
+import com.peekr.core.data.userKeyword.network.request.CreateUserKeywordRequest
+import com.peekr.core.data.userKeyword.network.request.PatchUserKeywordRequest
+import com.peekr.core.data.userKeyword.network.response.UserKeywordResponse
+import com.peekr.core.data.userKeyword.network.response.UserKeywordsResponse
 import com.peekr.core.domain.model.UserId
 import com.peekr.core.domain.model.UserKeywordId
 import javax.inject.Inject
