+++ core/data/src/main/java/com/peekr/core/data/userKeyword/network/response/UserKeywordsResponse.kt
@@ -1,4 +1,4 @@
-package com.peekr.core.data.userKeyword.model.response
+package com.peekr.core.data.userKeyword.network.response
 
 import com.peekr.core.domain.userKeyword.model.UserKeywords
 import com.squareup.moshi.JsonClass
