+++ core/data/src/test/java/com/peekr/core/data/file/network/FileNetworkDataSourceImplTest.kt
@@ -2,7 +2,7 @@ package com.peekr.core.data.file.network
 
 import com.peekr.core.common.logger.AppLogger
 import com.peekr.core.data.ServerTestRule
-import com.peekr.core.data.file.response.PresignedUrlResponse
+import com.peekr.core.data.file.network.response.PresignedUrlResponse
 import com.peekr.core.data.network.util.NetworkErrorType
 import com.peekr.core.data.network.util.NetworkResult
 import com.squareup.moshi.JsonDataException
