+++ core/data/src/test/java/com/peekr/core/data/userKeyword/repository/UserKeywordRepositoryImplTest.kt
@@ -3,13 +3,12 @@ package com.peekr.core.data.userKeyword.repository
 import com.peekr.core.data.network.util.NetworkErrorType
 import com.peekr.core.data.network.util.NetworkResult
 import com.peekr.core.data.network.util.toErrorType
-import com.peekr.core.data.userKeyword.model.request.CreateUserKeywordRequest
-import com.peekr.core.data.userKeyword.model.request.PatchUserKeywordRequest
-import com.peekr.core.data.userKeyword.model.response.UserKeywordResponse
-import com.peekr.core.data.userKeyword.model.response.UserKeywordsResponse
-import com.peekr.core.data.userKeyword.model.response.toDomainModel
 import com.peekr.core.data.userKeyword.network.UserKeywordDataSource
-import com.peekr.core.data.userKeyword.repository.UserKeywordRepositoryImpl
+import com.peekr.core.data.userKeyword.network.request.CreateUserKeywordRequest
+import com.peekr.core.data.userKeyword.network.request.PatchUserKeywordRequest
+import com.peekr.core.data.userKeyword.network.response.UserKeywordResponse
+import com.peekr.core.data.userKeyword.network.response.UserKeywordsResponse
+import com.peekr.core.data.userKeyword.network.response.toDomainModel
 import com.peekr.core.domain.model.KeywordId
 import com.peekr.core.domain.model.UserId
 import com.peekr.core.domain.model.UserKeywordId
